+++ query/query_test.go
@@ -46,7 +46,7 @@ func TestQuery(t *testing.T) {
 	t.Log(float64(bm.GetCardinality())/dur.Seconds(), "metrics/sec")
 
 	var name []byte
-	idx.MetricHashes(bm, func(u uint32, h histdb.Hash) bool {
+	idx.MetricHashes(bm, func(u uint64, h histdb.Hash) bool {
 		name, _ = idx.AppendMetricName(u, name[:0])
 		t.Logf("%-10d %x %s", u, h, name)
 		return true
